Cherry-picking EASY-1088: Prevent int overload by using a long.
and some corrections of spelling mistakes in the variable names. (tolarge -> toolarge)

+++ lib-deprecated/dans-lang/src/main/java/nl/knaw/dans/common/lang/service/exceptions/ZipFileLengthException.java
@@ -18,7 +18,7 @@ public class ZipFileLengthException extends FileSizeException {
     /*
      * Note that sizes in MegaBytes
      */
-    public ZipFileLengthException(int amount, int limit) {
+    public ZipFileLengthException(long amount, int limit) {
         super("The zip file exceeds the max length of " + limit + "megabytes");
         this.amount = amount;
         this.limit = limit;
